Include nginx.conf in files pushed to PaaS
Currently our manifest.yaml pushes only files built by middleman in the build
directory. This doesn't include the nginx.conf in the root of the repository.
This change moves it into the source files, this way it will be copied along
with other files into the build directory and correctly pushed to the paas.